Move parse unit tests to parse
