+++ src/parse/collection.rs
@@ -97,3 +97,113 @@ pub fn parse_expressions(it: &mut LexIterator) -> ParseResult<Vec<AST>> {
     })?;
     Ok(expressions)
 }
+
+#[cfg(test)]
+mod test {
+    use crate::lex::tokenize;
+    use crate::parse::ast::Node;
+    use crate::parse::parse_direct;
+
+    #[test]
+    fn tuple_empty_verify() {
+        let source = String::from("()");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let elements = match &statements.first().expect("script empty.").node {
+            Node::Tuple { elements } => elements,
+            _ => panic!("first element script was not tuple.")
+        };
+
+        assert_eq!(elements.len(), 0);
+    }
+
+    #[test]
+    fn tuple_single_is_expr_verify() {
+        let source = String::from("(a)");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let lit = match &statements.first().expect("script empty.").node {
+            Node::Id { lit } => lit,
+            _ => panic!("first element script was not tuple.")
+        };
+
+        assert_eq!(lit.as_str(), "a");
+    }
+
+    #[test]
+    fn tuple_multiple_verify() {
+        let source = String::from("(d, c)");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let elements = match &statements.first().expect("script empty.").node {
+            Node::Tuple { elements } => elements,
+            _ => panic!("first element script was not tuple.")
+        };
+
+        assert_eq!(elements.len(), 2);
+        assert_eq!(elements[0].node, Node::Id { lit: String::from("d") });
+        assert_eq!(elements[1].node, Node::Id { lit: String::from("c") });
+    }
+
+    #[test]
+    fn set_verify() {
+        let source = String::from("{a, b}");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let elements = match &statements.first().expect("script empty.").node {
+            Node::Set { elements } => elements,
+            _ => panic!("first element script was not set.")
+        };
+
+        assert_eq!(elements[0].node, Node::Id { lit: String::from("a") });
+        assert_eq!(elements[1].node, Node::Id { lit: String::from("b") });
+    }
+
+    #[test]
+    fn set_builder_verify() {
+        let source = String::from("{a | c, d}");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let (items, conditions) = match &statements.first().expect("script empty.").node {
+            Node::SetBuilder { item, conditions } => (item.clone(), conditions.clone()),
+            _ => panic!("first element script was not set builder.")
+        };
+
+        assert_eq!(items.node, Node::Id { lit: String::from("a") });
+
+        assert_eq!(conditions.len(), 2);
+        assert_eq!(conditions[0].node, Node::Id { lit: String::from("c") });
+        assert_eq!(conditions[1].node, Node::Id { lit: String::from("d") });
+    }
+
+    #[test]
+    fn list_verify() {
+        let source = String::from("[a, b]");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let elements = match &statements.first().expect("script empty.").node {
+            Node::List { elements } => elements,
+            _ => panic!("first element script was not list.")
+        };
+
+        assert_eq!(elements[0].node, Node::Id { lit: String::from("a") });
+        assert_eq!(elements[1].node, Node::Id { lit: String::from("b") });
+    }
+
+    #[test]
+    fn list_builder_verify() {
+        let source = String::from("[a | c, d]");
+        let statements = parse_direct(&tokenize(&source).unwrap()).unwrap();
+
+        let (items, conditions) = match &statements.first().expect("script empty.").node {
+            Node::ListBuilder { item, conditions } => (item.clone(), conditions.clone()),
+            _ => panic!("first element script was not list builder.")
+        };
+
+        assert_eq!(items.node, Node::Id { lit: String::from("a") });
+
+        assert_eq!(conditions.len(), 2);
+        assert_eq!(conditions[0].node, Node::Id { lit: String::from("c") });
+        assert_eq!(conditions[1].node, Node::Id { lit: String::from("d") });
+    }
+}
