+++ src/parse/definition.rs
@@ -216,3 +216,321 @@ fn parse_variable_def(it: &mut LexIterator) -> ParseResult {
     let id = it.parse(&parse_expression_type, "variable definition", &start)?;
     parse_variable_def_id(&id, it)
 }
+
+#[cfg(test)]
+mod test {
+    use crate::lex::tokenize;
+    use crate::parse::{parse, parse_direct};
+    use crate::parse::ast::Node;
+
+    macro_rules! unwrap_func_definition {
+    ($ast:expr) => {{
+        let definition = $ast.first().expect("script empty.").clone();
+        match definition.node {
+            Node::FunDef { id, pure, args, ret, raises, body, .. } =>
+                (pure, id, args, ret, raises, body),
+            other => panic!("Expected variabledef but was {:?}.", other)
+        }
+    }};
+}
+
+    macro_rules! unwrap_definition {
+    ($ast:expr) => {{
+        let definition = $ast.first().expect("script empty.").node.clone();
+        match definition {
+            Node::VariableDef { mutable, var, ty, expr, forward } =>
+                (mutable, var, ty, expr, forward),
+            other => panic!("Expected variabledef but was {:?}.", other)
+        }
+    }};
+}
+
+    #[test]
+    fn empty_definition_verify() {
+        let source = String::from("def a");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, _type, expression, forward) = unwrap_definition!(ast);
+
+        assert_eq!(mutable, true);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(_type, None);
+        assert_eq!(expression, None);
+        assert_eq!(forward, vec![]);
+    }
+
+    #[test]
+    fn definition_verify() {
+        let source = String::from("def a := 10");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        assert_eq!(mutable, true);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(ty, None);
+        assert_eq!(forward, vec![]);
+
+        match expression {
+            Some(expr_pos) => assert_eq!(expr_pos.node, Node::Int { lit: String::from("10") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn mutable_definition_verify() {
+        let source = String::from("def fin a := 10");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        assert_eq!(mutable, false);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(ty, None);
+        assert_eq!(forward, vec![]);
+
+        match expression {
+            Some(expr_pos) => assert_eq!(expr_pos.node, Node::Int { lit: String::from("10") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn private_definition_verify() {
+        let source = String::from("def a := 10");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        assert_eq!(mutable, true);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(ty, None);
+        assert_eq!(forward, vec![]);
+
+        match expression {
+            Some(expr_pos) => assert_eq!(expr_pos.node, Node::Int { lit: String::from("10") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn typed_definition_verify() {
+        let source = String::from("def a: Object := 10");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        let type_id = match ty {
+            Some(_type_pos) => match _type_pos.node {
+                Node::Type { id, generics: _ } => id,
+                other => panic!("Expected type but was: {:?}", other)
+            },
+            None => panic!("Expected type but was none.")
+        };
+        let expr = match expression {
+            Some(expr_pos) => expr_pos,
+            other => panic!("Unexpected expression: {:?}", other)
+        };
+
+        assert_eq!(mutable, true);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(forward, vec![]);
+        assert_eq!(expr.node, Node::Int { lit: String::from("10") });
+        assert_eq!(type_id.node, Node::Id { lit: String::from("Object") });
+    }
+
+    #[test]
+    fn forward_empty_definition_verify() {
+        let source = String::from("def a forward b, c");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        assert!(mutable);
+        assert_eq!(ty, None);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(expression, None);
+        assert_eq!(forward.len(), 2);
+        assert_eq!(forward[0].node, Node::Id { lit: String::from("b") });
+        assert_eq!(forward[1].node, Node::Id { lit: String::from("c") });
+    }
+
+    #[test]
+    fn forward_definition_verify() {
+        let source = String::from("def a := MyClass forward b, c");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (mutable, id, ty, expression, forward) = unwrap_definition!(ast);
+
+        assert!(mutable, true);
+        assert_eq!(ty, None);
+        assert_eq!(id.node, Node::Id { lit: String::from("a") });
+        assert_eq!(expression.unwrap().node, Node::Id { lit: String::from("MyClass") });
+        assert_eq!(forward.len(), 2);
+        assert_eq!(forward[0].node, Node::Id { lit: String::from("b") });
+        assert_eq!(forward[1].node, Node::Id { lit: String::from("c") });
+    }
+
+    #[test]
+    fn function_definition_verify() {
+        let source = String::from("def f(fin b: Something, vararg c) => d");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (pure, id, fun_args, ret, raises, body) = unwrap_func_definition!(ast);
+
+        assert!(!pure);
+        assert_eq!(id.node, Node::Id { lit: String::from("f") });
+        assert_eq!(fun_args.len(), 2);
+        assert_eq!(ret, None);
+        assert_eq!(raises, vec![]);
+
+        match body {
+            Some(body) => assert_eq!(body.node, Node::Id { lit: String::from("d") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+
+        match (&fun_args[0].node, &fun_args[1].node) {
+            (
+                Node::FunArg { vararg: v1, var: id1, mutable: mut1, ty: ty1, default: d1 },
+                Node::FunArg { vararg: v2, var: id2, mutable: mut2, ty: ty2, default: d2 }
+            ) => {
+                assert_eq!(v1.clone(), false);
+                assert_eq!(v2.clone(), true);
+
+                assert_eq!(id1.node, Node::Id { lit: String::from("b") });
+                assert_eq!(id2.node, Node::Id { lit: String::from("c") });
+
+                assert!(!mut1);
+                assert!(mut2);
+
+                match ty1.clone().unwrap().node {
+                    Node::Type { id, generics } => {
+                        assert_eq!(id.node, Node::Id { lit: String::from("Something") });
+                        assert_eq!(generics.len(), 0);
+                    }
+                    other => panic!("Expected type for first argument: {:?}", other)
+                }
+                assert_eq!(ty2.clone(), None);
+
+                assert_eq!(d1.clone(), None);
+                assert_eq!(d2.clone(), None);
+            }
+            other => panic!("Expected two fun args: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn function_no_args_definition_verify() {
+        let source = String::from("def f() => d");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (pure, id, args, ret, _, body) = unwrap_func_definition!(ast);
+
+        assert!(!pure);
+        assert_eq!(id.node, Node::Id { lit: String::from("f") });
+        assert_eq!(args.len(), 0);
+        assert_eq!(ret, None);
+
+        match body {
+            Some(body) => assert_eq!(body.node, Node::Id { lit: String::from("d") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn function_pure_definition_verify() {
+        let source = String::from("def pure f() => d");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (pure, id, args, ret, _, body) = unwrap_func_definition!(ast);
+
+        assert!(pure);
+        assert_eq!(id.node, Node::Id { lit: String::from("f") });
+        assert_eq!(args.len(), 0);
+        assert_eq!(ret, None);
+
+        match body {
+            Some(body) => assert_eq!(body.node, Node::Id { lit: String::from("d") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn function_definition_with_literal_verify() {
+        let source = String::from("def f(x, vararg b: Something) => d");
+        let ast = parse_direct(&tokenize(&source).unwrap()).unwrap();
+        let (pure, id, fun_args, ret, _, body) = unwrap_func_definition!(ast);
+
+        assert!(!pure);
+        assert_eq!(id.node, Node::Id { lit: String::from("f") });
+        assert_eq!(fun_args.len(), 2);
+        assert_eq!(ret, None);
+
+        match body {
+            Some(body) => assert_eq!(body.node, Node::Id { lit: String::from("d") }),
+            other => panic!("Unexpected expression: {:?}", other)
+        }
+
+        match (&fun_args[0].node, &fun_args[1].node) {
+            (
+                Node::FunArg { vararg: v1, var: id1, mutable: mut1, ty: ty1, default: d1 },
+                Node::FunArg { vararg: v2, var: id2, mutable: mut2, ty: ty2, default: d2 }
+            ) => {
+                assert!(!v1.clone());
+                assert!(v2.clone());
+
+                assert!(mut1.clone());
+                assert!(mut2.clone());
+
+                assert_eq!(id1.node, Node::Id { lit: String::from("x") });
+                assert_eq!(id2.node, Node::Id { lit: String::from("b") });
+
+                assert_eq!(ty1.clone(), None);
+                match ty2.clone().unwrap().node {
+                    Node::Type { id, generics } => {
+                        assert_eq!(id.node, Node::Id { lit: String::from("Something") });
+                        assert_eq!(generics.len(), 0);
+                    }
+                    other => panic!("Expected type for first argument: {:?}", other)
+                }
+
+                assert_eq!(d1.clone(), None);
+                assert_eq!(d2.clone(), None);
+            }
+            other => panic!("Expected two fun args: {:?}", other)
+        }
+    }
+
+    #[test]
+    fn def_mut_private_wrong_order() {
+        let source = String::from("def mut private a ");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+    #[test]
+    fn def_missing_id() {
+        let source = String::from("def");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+    #[test]
+    fn def_fun_no_closing_brack() {
+        let source = String::from("def f(a");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+    #[test]
+    fn def_fun_missing_arrow() {
+        let source = String::from("def f(a) a * 10");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+    #[test]
+    fn def_fun_missing_brackets() {
+        let source = String::from("def f => print a");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+
+    #[test]
+    fn handle_no_branches() {
+        let source = String::from("def a handle");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+
+    #[test]
+    fn handle_no_indentation() {
+        let source = String::from("def a handle\nerr: Err => b");
+        parse(&tokenize(&source).unwrap()).unwrap_err();
+    }
+}
