+++ tests/parse/invalid/control_flow.rs
@@ -1,69 +1,22 @@
 use mamba::lex::tokenize;
+use mamba::parse::parse;
 
-use crate::parse::util::parse_direct;
+use crate::common::resource_content;
 
 #[test]
-fn for_missing_do() {
-    let source = String::from("for a in c d");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
+fn for_statements() {
+    let source = resource_content(true, &["control_flow"], "for_statements.mamba");
+    parse(&tokenize(&source).unwrap()).unwrap();
 }
 
 #[test]
-fn for_missing_body() {
-    let source = String::from("for a in c");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
+fn if_stmt() {
+    let source = resource_content(true, &["control_flow"], "if.mamba");
+    assert!(parse(&tokenize(&source).unwrap()).is_ok());
 }
 
 #[test]
-fn if_missing_then() {
-    let source = String::from("if a b");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn if_missing_body() {
-    let source = String::from("if a then");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn if_then_missing_body() {
-    let source = String::from("if a then b else");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn match_missing_condition() {
-    let source = String::from("match\n    a => b");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn match_missing_arms() {
-    let source = String::from("match a with\n    ");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn match_missing_arms_no_newline() {
-    let source = String::from("match a");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn while_missing_condition() {
-    let source = String::from("while do b");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn while_missing_body() {
-    let source = String::from("while a do");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
-}
-
-#[test]
-fn while_missing_do() {
-    let source = String::from("while a b");
-    parse_direct(&tokenize(&source).unwrap()).unwrap_err();
+fn match_statements() {
+    let source = resource_content(true, &["control_flow"], "match.mamba");
+    parse(&tokenize(&source).unwrap()).unwrap();
 }
