+++ tests/parse/invalid/mod.rs
@@ -1,7 +1,3 @@
 pub mod compound;
-pub mod control_flow;
-pub mod definition;
-pub mod error;
-pub mod expression_and_statement;
-pub mod function;
-pub mod operation;
+mod control_flow;
+mod definition;
