+++ tests/parse/mod.rs
@@ -5,7 +5,6 @@ use crate::common::*;
 
 pub mod invalid;
 pub mod valid;
-mod util;
 
 #[test]
 fn parse_empty_file() {
