+++ tests/parse/valid/mod.rs
@@ -1,9 +1,5 @@
-pub mod call;
 pub mod class;
 pub mod collection;
-pub mod control_flow;
-pub mod definition;
+mod control_flow;
 pub mod error;
-pub mod expression_and_statement;
 pub mod function;
-pub mod operation;
